+++ src/main/java/com/vietnam/lottery/business/basicIndicators/response/ProbabilityResponse.java
@@ -4,7 +4,6 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 public class ProbabilityResponse implements Serializable {
@@ -15,8 +14,8 @@ public class ProbabilityResponse implements Serializable {
     private String name;
 
     @ApiModelProperty(value = "金额")
-    private BigDecimal amount;
+    private Long amount;
 
     @ApiModelProperty(value = "中奖概率")
-    private BigDecimal probability;
+    private Long probability;
 }
