+++ src/main/java/com/vietnam/lottery/business/unpackRedPackets/request/UnPackAddRequest.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 public class UnPackAddRequest implements Serializable {
@@ -26,7 +25,7 @@ public class UnPackAddRequest implements Serializable {
 
     @ApiModelProperty(value = "中奖概率")
     @NotNull(message = "中奖概率不能为空")
-    private BigDecimal probability;
+    private Long probability;
 
     @ApiModelProperty(hidden = true)
     private Long createBy;
